+++ components/hooks/useRenderQuestionField.tsx
@@ -70,11 +70,13 @@ export interface RenderQuestionFieldProps {
   textAreaProps?: {
     content: string;
     setContent?: (newContent: string) => void;
+    handleTextAreaChange?: () => void;
   };
 
   selectBoxProps?: {
     selectedSelectValue?: string | undefined;
     setSelectedSelectValue?: (val: string | undefined) => void;
+    handleSelectChange?: (value: string) => void;
   };
 
   multiSelectBoxProps?: {
@@ -97,13 +99,13 @@ export interface RenderQuestionFieldProps {
   };
 
   numberProps?: {
-    inputValue: number | null;
-    setInputValue: (val: number | null) => void
+    numberValue: number | null;
+    handleNumberChange: (value: number) => void;
   };
 
   numberRangeProps?: {
     numberRange: { startNumber: number | null; endNumber: number | null };
-    handleNumberChange: (key: string, val: number | null) => void
+    handleNumberRangeChange: (key: string, val: number | null) => void
   };
 
   checkBoxProps?: {
@@ -113,7 +115,7 @@ export interface RenderQuestionFieldProps {
 
   currencyProps?: {
     inputCurrencyValue: number | null;
-    setInputCurrencyValue: (val: number | null) => void
+    handleCurrencyChange: (value: number | null) => void;
   };
 
   booleanProps?: {
@@ -199,6 +201,7 @@ export function useRenderQuestionField({
                   parsedQuestion={parsed}
                   selectedSelectValue={selectBoxProps.selectedSelectValue}
                   setSelectedSelectValue={selectBoxProps.setSelectedSelectValue}
+                  handleSelectChange={selectBoxProps.handleSelectChange}
               />
           );
         }
@@ -245,6 +248,7 @@ export function useRenderQuestionField({
             id="question-text-editor"
             content={textAreaProps?.content ?? ''}
             setContent={(value) => textAreaProps?.setContent && textAreaProps?.setContent(value)}
+            onChange={textAreaProps?.handleTextAreaChange}
           />
         );
       }
@@ -285,9 +289,9 @@ export function useRenderQuestionField({
         return (
           <NumberComponent
             label="number"
-            value={numberProps?.inputValue ?? undefined
+            value={numberProps?.numberValue ?? undefined
             }
-            onChange={numberProps?.setInputValue}
+            onChange={numberProps?.handleNumberChange}
             placeholder="number"
             minValue={minValue}
             {...(typeof maxValue === 'number' ? { maxValue } : {})}
@@ -303,7 +307,7 @@ export function useRenderQuestionField({
           <NumberRangeQuestionComponent
             parsedQuestion={parsed}
             numberRange={numberRangeProps?.numberRange}
-            handleNumberChange={numberRangeProps?.handleNumberChange}
+            handleNumberChange={numberRangeProps?.handleNumberRangeChange}
             startPlaceholder="start"
             endPlaceholder="end"
           />
@@ -317,7 +321,7 @@ export function useRenderQuestionField({
           <CurrencyQuestionComponent
             parsedQuestion={parsed}
             inputCurrencyValue={currencyProps?.inputCurrencyValue}
-            setInputCurrencyValue={currencyProps?.setInputCurrencyValue}
+            handleCurrencyChange={currencyProps?.handleCurrencyChange}
           />
         );
       }